Fix snake case formatting
Check the previous character to determine whether an underscore should be added before an uppercase letter. It only adds an underscore if the previous character is not an uppercase letter, which ensures that acronyms and abbreviations are properly handled. Also handle the scenario where there are already underscores in the column name.